refactor: add separate components
